Account: cancel in-flight network requests on account state transitions
