+++ components/brave_account/DEPS
@@ -1,4 +1,5 @@
 include_rules = [
+  "+absl/functional",
   "+components/grit/brave_components_resources.h",
   "+components/grit/brave_components_strings.h",
   "+components/grit/brave_components_webui_strings.h",
